+++ src/main/kotlin/ara/analysis/LocalDeclarationAnalysis.kt
@@ -30,10 +30,13 @@ class LocalDeclarationAnalysis(private val program: Syntax.Program) : Analysis<U
         is Syntax.TypedStorage ->
             declare(expression.storage)
 
-        is Syntax.MemberAccess ->
+        is Syntax.MemberAccess -> // TODO: Perhaps we should not allow the declaration of variables as part of a member access...
             declare(expression.storage)
 
         is Syntax.IntegerLiteral ->
+            Unit
+
+        is Syntax.StructureLiteral ->
             Unit
     }