+++ src/main/kotlin/ara/storage/ResourceAllocation.kt
@@ -4,24 +4,30 @@ import ara.syntax.Syntax
 
 object ResourceAllocation {
 
-    fun Syntax.ResourceExpression.asResourcePath(): ResourcePath? = when (this) {
-        is Syntax.MemberAccess -> {
-            val result = this.storage.asResourcePath()
-            result?.appended(this.member)
-        }
-
+    fun Syntax.Storage.asResourcePath(): ResourcePath = when (this) {
         is Syntax.NamedStorage ->
             ResourcePath.ofIdentifier(this.name)
 
         is Syntax.TypedStorage ->
             this.storage.asResourcePath()
 
+        is Syntax.MemberAccess ->
+            this.storage.asResourcePath().appended(this.member)
+    }
+
+    fun Syntax.ResourceExpression.asResourcePaths(): Collection<ResourcePath> = when (this) {
+        is Syntax.Storage ->
+            listOf(this.asResourcePath())
+
         is Syntax.IntegerLiteral ->
-            null
+            emptyList()
+
+        is Syntax.StructureLiteral ->
+            this.members.flatMap { it.value.asResourcePaths() }
     }
 
     fun Syntax.Instruction.variablesCreated(): Collection<ResourcePath> =
-        resourcesCreated().mapNotNull { it.asResourcePath() }
+        this.resourcesCreated().flatMap { it.asResourcePaths() }
 
     fun Syntax.Instruction.resourcesCreated(): Collection<Syntax.ResourceExpression> = when (this) {
         is Syntax.Assignment ->
@@ -35,7 +41,7 @@ object ResourceAllocation {
     }
 
     fun Syntax.Instruction.variablesDestroyed(): Collection<ResourcePath> =
-        this.resourcesDestroyed().mapNotNull { it.asResourcePath() }
+        this.resourcesDestroyed().flatMap { it.asResourcePaths() }
 
     fun Syntax.Instruction.resourcesDestroyed(): Collection<Syntax.ResourceExpression> = when (this) {
         is Syntax.Assignment ->
@@ -50,15 +56,15 @@ object ResourceAllocation {
 
     fun Syntax.ArithmeticExpression.asResourcePaths(): Collection<ResourcePath> = when (this) {
         is Syntax.ArithmeticBinary ->
-            listOfNotNull(this.lhs.asResourcePath(), this.rhs.asResourcePath())
+            this.lhs.asResourcePaths() + this.rhs.asResourcePaths()
 
         is Syntax.ArithmeticValue ->
-            listOfNotNull(this.value.asResourcePath())
+            this.value.asResourcePaths()
     }
 
     fun Syntax.ConditionalExpression.asResourcePaths(): Collection<ResourcePath> = when (this) {
         is Syntax.ComparativeBinary ->
-            listOfNotNull(this.lhs.asResourcePath(), this.rhs.asResourcePath())
+            this.lhs.asResourcePaths() + this.rhs.asResourcePaths()
     }
 
     fun Syntax.Instruction.variablesUsed(): Collection<ResourcePath> = when (this) {