+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Structure literals to create structures without manually implemented constructors.
+
+### Fixed
+
+- Bug in liveness analysis where variables of a structure type could not be initialized or finalized.
+
 ## [indev-1] - 2023-08-06
 
 ### Added